+++ src/components/Card/Card.tsx
@@ -1,9 +1,13 @@
 import React from "react";
-import { Card as BlueprintCard, CardProps as BlueprintCardProps } from "@blueprintjs/core";
+import {
+    Card as BlueprintCard,
+    CardProps as BlueprintCardProps,
+    Elevation as BlueprintCardElevation,
+} from "@blueprintjs/core";
 
 import { CLASSPREFIX as eccgui } from "../../configuration/constants";
 
-export interface CardProps extends BlueprintCardProps {
+export interface CardProps extends Omit<BlueprintCardProps, "elevation"> {
     /**
      * `<Card />` element is included in DOM as simple `div` element.
      * By default it is a HTML `section`.
@@ -17,6 +21,12 @@ export interface CardProps extends BlueprintCardProps {
      * Background color is slightly altered to differ card display from other cards.
      */
     elevated?: boolean;
+    /**
+     * Controls the intensity of the drop shadow beneath the card.
+     * At elevation `0`, no drop shadow is applied.
+     * At elevation `-1`, the card is even borderless.
+     */
+    elevation?: -1 | BlueprintCardElevation;
     /**
      * When card (or its children) get focus the card is scrolled into the viewport.
      * Property value defined which part of the card is always scrolled in, this may important when the card is larger than the viewport.
@@ -35,7 +45,7 @@ export interface CardProps extends BlueprintCardProps {
 export const Card = ({
     children,
     className = "",
-    elevation = 1,
+    elevation = -1,
     isOnlyLayout = false,
     fullHeight = false,
     elevated = false,
@@ -48,6 +58,7 @@ export const Card = ({
         ? {
               tabIndex: 0,
               onFocus: (e: any) => {
+                  // FIXME: we should not have any hard relations to apps that using this lib
                   const el = e.target.closest(".diapp-iframewindow__content");
                   setTimeout(() => {
                       if (el)
@@ -68,9 +79,10 @@ export const Card = ({
                 (elevated ? ` ${eccgui}-card--elevated` : "") +
                 (scrollinOnFocus ? ` ${eccgui}-card--scrollonfocus` : "") +
                 (whitespaceAmount !== "medium" ? ` ${eccgui}-card--whitespace-${whitespaceAmount}` : "") +
+                (elevation < 0 ? ` ${eccgui}-card--whitespace-borderless` : "") +
                 (className ? ` ${className}` : "")
             }
-            elevation={elevation}
+            elevation={Math.max(0, elevation) as BlueprintCardElevation}
             interactive={otherProps.onClick ? true : interactive}
             {...scrollIn}
             {...otherProps}